// frontend/src/app/generated/model/modelsAuditLog.ts
/**
 * Homelab API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface ModelsAuditLog {
  /**
   * CREATE, UPDATE, DELETE, or HTTP Method
   */
  action?: string;
  id?: string;
  ipAddress?: string;
  /**
   * Detailed description
   */
  message?: string;
  /**
   * e.g., \"DNS/Record\", \"RBAC/ServiceAccount\"
   */
  resource?: string;
  /**
   * Success or Failed
   */
  status?: string;
  /**
   * User or SA name (e.g., \"root\" or \"sa-name\")
   */
  subject?: string;
  /**
   * ID or name of the resource operated on
   */
  targetId?: string;
  /**
   * RFC3339 format
   */
  timestamp?: string;
  userAgent?: string;
}